move name-importing to PostRector
